Add LICENSE file (#31)
I want to use the
[devcontainers/ci](https://github.com/devcontainers/ci/blob/main/README.md)
action to run trunk check inside a dev container, but the official Trunk
action can only be run on a standard GitHub Actions runner. As a result,
I need to manually install Trunk and its dependencies, and invoke Trunk
in a manner similar to the
[pull_request.sh](https://github.com/trunk-io/trunk-action/blob/main/pull_request.sh)
script. I am creating a custom `trunk-check.sh` script to handle this
process, and I am using this repository as a reference for my
implementation.

I want to credit you properly, with copyright and license information
included. However, since this repository does not have a `LICENSE` file,
the code is not open source (cf. [Choose a License: No
License](https://choosealicense.com/no-permission/)). However, I noticed
that you seem to use the MIT license in your other repositories, so this
pull request adds the MIT license to this repository.