Fix repo tests to error if trunk-check doesn't run (#88)
The repo tests now return an error instead of success if
landing-state.json does not exist. Additionally, the job will show as
failed if any individual job fails (and it still runs all tests, rather
than quitting on the first fail).

I also bumped all the refs for the repos to latest main.

However, several repos fail the test, all for different reasons, and I'm
not quite sure how to handle that:
 - vuejs/core fails because pnpm version is too new
- shopify/draggable fails because webpack doesn't work with newest npm:
https://github.com/webpack/webpack/issues/14532
- sheldonhull/sheldonhull.hugo fails because yarn install would update
the yarn lockfile
- pallets/flask fails because of a linter failure in an invalid html
file

I'm PRing this now to get eyes on it, despite likely needing to fix the
issues with the repos before merging.