fix: actually install node during setup (#91)
We have never passed a node version into the setup-node action, making
that action a no-op. This worked previously because the github
`ubuntu-latest` runners have node installed by default. However, after
switching over to our self-hosted runners that do not have node
installed, the action started failing.

This PR adds logic to detect files that have the node version, and pass
them into the setup_node action to be parsed. It also adds
shopify/draggable back to the repo_tests, as it was failing because the
node version was too new.