fix: cache linters/formatters correctly (#93)
Because all of the caching is done in the .trunk repo, the caches for
different repos in the same org stomp on each other. This introduces
different keys for different repos. This can be seen in the [repo_tests
running on this
PR](https://github.com/trunk-io/trunk-action/actions/runs/4887481583/jobs/8724116197?pr=93).

Additionally, this PR adds caching for the `node_modules` folder,
following the same caching logic.

This has an associated PR in trunk-io/trunk,
[#8408](https://github.com/trunk-io/trunk/pull/8408).

---------

Co-authored-by: Samuel Lijin <sam@trunk.io>