feat: finalize via CheckService instead of the GH API
Going via CheckService allows us to no longer have to worry about the
GH token expiring too soon; in exchange, we have to accept that there
are CLI->CheckService gaps we may not have coverage for.

This will, however, ensure coverage of the case where something in
normal "trunk check" goes wrong, which is the primary case that we're
concerned about.