fix: set up inputs correctly (#138)
#127 broke check-on-PRs again

* so we know we really need to prio adding test coverage for the
CheckService callpath in trunk-action
* the reason that PR broke this is because we're not in a git repo until
actions/checkout runs, so `.git` does not yet exist, and therefore we
can't write to `.git/info/exclude`
* in addition, by bootstrapping `.trunk`, we're putting stuff in the
current repository, which `actions/checkout` will have to stomp

This changes things to bank on the assumption that `.trunk-action` does
not exist and that it's safe to always clean it up. Also make two other
housekeeping changes:

* rename mask-payload to input-setup, which is more accurate, and
* bump the z-shell/wiki ref in repo_tests since stylelint keeps
segfaulting on the current ref in repo_tests.