fix: move input setup to inline bash (#145)
Because of path shenanigans when running the trunk-action in a
container, the input-setup action wasn't getting found by node. To avoid
this, I've rewritten the logic in bash. This PR also fixes TRUNK-7633
and adds tests for running the action in a container, so we don't get
blindsided by something like this again.

Workflow dispatch code path was tested manually with the services cli
tool
[here](https://github.com/trunk-io/.trunk-internal/actions/runs/5350157016/jobs/9702379590).

-----
Some context on the things I tried prior to this:
- I tried using `$GITHUB_ACTION_PATH` instead of `${{
github.action_path}}` to find the action, as the environment variable is
set correctly, but the context variable is not. However, using the env
var, github fails to find the action at all. Using the context variable,
github finds the `action.yml` and `mask.js` files, but passes an
incorrect path into node to run.
- I considered calling the subaction directly with `uses:
trunk-io/trunk-action/input-setup@v1` - however, this would potentially
introduce a version mismatch between the version of the action being run
and the version of the subaction being wrong
- I considered simply calling `node mask.js` inline in the composite
action, but didn't want to add a setup node action that would
potentially conflict with another node install
- I tried to use templating in the main filename javascript action to
pass in the correct path - while you can use templating in that field,
you cannot access the `env` or `input` contexts, making it impossible to
set the field correctly.