fix: attempt to install specified pnpm version (#243)
Currently, we only install the latest pnpm version, which fails if the
user has specified a pnpm version under `packageManager` in their
package.json. This PR changes that to attempt to parse the version out
of the package.json before defaulting to latest.

The `pnpm/action-setup` action does have logic for parsing the version
out of the package.json, but we cannot provide both an version and a
version in the package.json, or it will fail ([source
link](https://github.com/pnpm/action-setup/blob/master/src/install-pnpm/run.ts#L64)).

Tested manually, both with and without the `packageManager` field.