feat(upgrade): update create-pull-request to v7 and add sign-commits input parameter (#263)
I want to use automated trunk updates in a repository where all commits
need to be signed.

So I thought about adding it to the github action because the
`create-pull-request` action that is used here has support for signed
commits using GitHub API starting with v7.

So I upgraded `create-pull-request` to v7 and added a new input value to
enable signing.

I am using my forked action currently and its working fine but I would
like to contribute this to you :)

---------

Co-authored-by: Tyler Jang <tyler@trunk.io>